# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: SaaS Onboarding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/create-subscription:
    post:
      operationId: createMpSubscription
      tags: [business]
      summary: Create a Mercado Pago subscription (returns init_point URL)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMpSubscriptionInput"
      responses:
        "200":
          description: Subscription init point returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateMpSubscriptionResult"
        "401":
          description: Unauthorized
  /billing/webhook:
    post:
      operationId: receiveMpWebhook
      tags: [business]
      summary: Mercado Pago webhook (public — HMAC-verified)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MpWebhookNotification"
      responses:
        "200":
          description: Accepted
  /billing/status:
    get:
      operationId: getBillingStatus
      tags: [business]
      summary: Get current billing status for authenticated tenant
      responses:
        "200":
          description: Billing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingStatus"
        "401":
          description: Unauthorized
  /billing/cancel:
    post:
      operationId: cancelBillingSubscription
      tags: [business]
      summary: Cancel current subscription
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingStatus"
        "401":
          description: Unauthorized
  /billing/payment-history:
    get:
      operationId: getPaymentHistory
      tags: [business]
      summary: Get payment event history for authenticated tenant
      responses:
        "200":
          description: Payment events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentEvent"
        "401":
          description: Unauthorized
components:
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        plan:
          type: string
        status:
          type: string
          enum: [active, cancelled, expired, trialing, past_due, paused]
        billingCycle:
          type: string
          enum: [monthly, yearly]
        amount:
          type: ["string", "null"]
        currency:
          type: string
        startedAt:
          type: string
        expiresAt:
          type: ["string", "null"]
        currentPeriodStart:
          type: ["string", "null"]
        currentPeriodEnd:
          type: ["string", "null"]
        mercadoPagoSubscriptionId:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        nextBillingAt:
          type: ["string", "null"]
        failedPaymentsCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - plan
        - status
        - billingCycle
        - currency
        - startedAt
        - failedPaymentsCount
        - createdAt
    CreateMpSubscriptionInput:
      type: object
      properties:
        plan:
          type: string
          enum: [pro, premium]
        billingCycle:
          type: string
          enum: [monthly, yearly]
        paymentMethod:
          type: string
          enum: [credit_card, pix]
        backUrl:
          type: string
      required:
        - plan
        - billingCycle
    CreateMpSubscriptionResult:
      type: object
      properties:
        initPoint:
          type: string
        subscriptionId:
          type: string
        preapprovalId:
          type: ["string", "null"]
      required:
        - initPoint
        - subscriptionId
    BillingStatus:
      type: object
      properties:
        subscription:
          allOf:
            - $ref: "#/components/schemas/Subscription"
          nullable: true
        trialEndsAt:
          type: ["string", "null"]
        trialDaysLeft:
          type: ["integer", "null"]
        isTrialing:
          type: boolean
        isActive:
          type: boolean
        isPastDue:
          type: boolean
        planLocked:
          type: boolean
      required:
        - isTrialing
        - isActive
        - isPastDue
        - planLocked
    PaymentEvent:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        subscriptionId:
          type: ["string", "null"]
        mercadoPagoPaymentId:
          type: ["string", "null"]
        status:
          type: string
          enum: [approved, pending, in_process, rejected, cancelled, refunded, charged_back]
        paymentMethod:
          type: ["string", "null"]
        amount:
          type: string
        currency:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - status
        - amount
        - currency
        - createdAt
    MpWebhookNotification:
      type: object
      properties:
        id:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        action:
          type: ["string", "null"]
        data:
          type: object
          additionalProperties: true
...